add comment

+++ src/main/java/com/ry/community/controller/PublishController.java
@@ -27,7 +27,7 @@ public class PublishController {
 
     @GetMapping("/publish/{id}")
     public String edit(
-            @PathVariable(name = "id") Integer id,
+            @PathVariable(name = "id") Long id,
             Model model
     ) {
         QuestionDTO questuin = questionService.findQuestionDTOById(id);
@@ -49,7 +49,7 @@ public class PublishController {
             @RequestParam(value = "title", required = false) String title,
             @RequestParam(value = "description", required = false) String description,
             @RequestParam(value = "tags", required = false) String tags,
-            @RequestParam(value = "id", required = false) Integer id,
+            @RequestParam(value = "id", required = false) Long id,
             HttpServletRequest request,
             Model model
     ) {
